Numerical snow layer issue
Removed the numerical values for snow layers
Using things like "1 thick snow layer" wasn't parsing the number as part of the snow layer but rather as the amount of items

Also note... I went thru the rest of the alias files. A bunch of numbers that did not need to be removed as the number is WITHIN the alias not at the beginning .... example "level 1 something" .... I did some testing, and because the number is inside the alias it seems to work properly.