Update alias for the 1.16 archery target block
This block's aliases were conflicting with the `target` expression used to retrieve the player's targeted entity, and making `block` not be optional wouldn't fix it as it would then conflict with the `target block` expression. We need to add words to distinguish this as the in-game block rather than it being associated with a player or entity's raytrace-targeted object.

Tried to add multiple sensible alias options to make it as intuitive as possible, but this is an unfortunate necessary evil.